added hooks folder and added file with objcts in it
